requirements.txt
django
gunicorn
whitenoise
psycopg2-binary
requests
urllib3
pillow
django-livereload-server
django-cors-headers
djangorestframework